Morse Coder Part 14 : Settings Page
Added a Settings View and View Model.

Added relevant MVVM Light plumbing to the ViewModelLocator

+++ MorseCoder/MorseCoderSettings.cs
@@ -42,7 +42,7 @@ namespace MorseCoder
             }
             set
             {
-                localSettings.Values[_directionKey] = value;
+                localSettings.Values[_directionKey] = (int)value;
             }
         }
     }
